requirements.txt
TensorFlow==2.10.0
Keras==2.10.0
numpy==1.21.6
matplotlib==3.5.1
scikit-learn==1.1.2
seaborn==0.11.2
Pillow==9.2.0
opencv-python==4.6.0.66
pandas==1.4.4
json5==0.9.6
torch>=1.13.0
torchvision>=0.14.0